test(dedupe): consolidate risk_manager tests (12 → 5 files)
Group by concern:
- test_risk_manager.py: init, stubs, ValidationError, risk_metrics
- test_risk_manager_daily.py: daily_pnl, reset_daily_tracking, mark_staleness
- test_risk_manager_validation.py: pre_trade_validate, workflows
- test_risk_manager_volatility.py: circuit_breaker, VolatilityCheckOutcome
- test_risk_manager_protection.py: liquidation_buffer, reduce_only_mode

Cluster: cebae704e17e
