+++ tests/_triage/dedupe_triage.yaml
@@ -1,13 +1,43 @@
 # Test Deduplication Triage
 #
-# This file stores human triage decisions for clusters in:
-#   tests/_triage/dedupe_candidates.yaml
-#
 # States:
 #   accepted: worth doing (keep surfaced)
 #   rejected: false positive / not worth doing
 #   deferred: maybe later (hide from suggestions)
 #
 version: 1
-generated_at: null
-clusters: {}
+generated_at: '2026-01-21T01:13:28.459963+00:00'
+clusters:
+  72948f9bf0a3:
+    state: rejected
+    owner: claude
+    updated: '2026-01-21'
+    reason: 'False positive: source_fanout based on shared models import. Files test
+      distinct functionality (auth, system, REST, market data) - 1005 lines total
+      would require 5+ files anyway.'
+  ce4068fdc367:
+    state: rejected
+    owner: claude
+    updated: '2026-01-21'
+    reason: 'False positive: source_fanout based on shared utilities import. Files
+      test distinct functionality (PnL, positions, REST, perps, catalog) - unrelated
+      concerns.'
+  7a9b6155b2a7:
+    state: rejected
+    owner: claude
+    updated: '2026-01-21'
+    reason: 'False positive: source_fanout based on validation import. Tests distinct
+      rule types (boolean, numeric, collection, string) in separate modules - good
+      organization.'
+  0bdbfcad487d:
+    state: rejected
+    owner: claude
+    updated: '2026-01-21'
+    reason: 'False positive: source_fanout mixing signals tests (order_flow, orderbook,
+      spread, vwap) with optimize runner - distinct modules, good organization.'
+  1707c929251f:
+    state: rejected
+    owner: claude
+    updated: '2026-01-21'
+    reason: 'False positive: source_fanout mixing app container test with secrets
+      manager tests - completely unrelated modules.'
