+++ tests/unit/gpt_trader/features/live_trade/test_risk_manager_validation.py
@@ -1,4 +1,4 @@
-"""Tests for LiveRiskManager.pre_trade_validate."""
+"""Tests for LiveRiskManager pre-trade validation and workflows."""
 
 from __future__ import annotations
 
@@ -22,7 +22,7 @@ class TestPreTradeValidate:
     """Tests for pre_trade_validate method."""
 
     def test_no_config_passes(self) -> None:
-        """Test validation passes when no config."""
+        """Should pass validation when no config."""
         manager = LiveRiskManager()
 
         result = manager.pre_trade_validate(
@@ -37,7 +37,7 @@ class TestPreTradeValidate:
         assert result is None
 
     def test_leverage_within_limit(self) -> None:
-        """Test validation passes when leverage within limit."""
+        """Should pass when leverage within limit."""
         config = MockConfig(max_leverage=Decimal("10"))
         manager = LiveRiskManager(config=config)
 
@@ -54,7 +54,7 @@ class TestPreTradeValidate:
         assert result is None
 
     def test_leverage_exceeds_limit_raises(self) -> None:
-        """Test validation raises when leverage exceeds limit."""
+        """Should raise when leverage exceeds limit."""
         config = MockConfig(max_leverage=Decimal("5"))
         manager = LiveRiskManager(config=config)
 
@@ -71,11 +71,10 @@ class TestPreTradeValidate:
             )
 
     def test_zero_equity_skips_validation(self) -> None:
-        """Test validation skipped when equity is zero."""
+        """Should skip validation when equity is zero."""
         config = MockConfig(max_leverage=Decimal("5"))
         manager = LiveRiskManager(config=config)
 
-        # Should not raise even with high notional
         result = manager.pre_trade_validate(
             symbol="BTC-USD",
             side="buy",
@@ -88,11 +87,10 @@ class TestPreTradeValidate:
         assert result is None
 
     def test_negative_equity_skips_validation(self) -> None:
-        """Test validation skipped when equity is negative."""
+        """Should skip validation when equity is negative."""
         config = MockConfig(max_leverage=Decimal("5"))
         manager = LiveRiskManager(config=config)
 
-        # Negative equity means the condition equity > 0 fails
         result = manager.pre_trade_validate(
             symbol="BTC-USD",
             side="buy",
@@ -103,3 +101,85 @@ class TestPreTradeValidate:
             current_positions={},
         )
         assert result is None
+
+
+class TestLiveRiskManagerWorkflows:
+    """Workflow tests for LiveRiskManager."""
+
+    def test_daily_workflow(self) -> None:
+        """Should handle typical daily workflow."""
+        config = MockConfig(
+            max_leverage=Decimal("10"),
+            daily_loss_limit_pct=Decimal("0.05"),
+            min_liquidation_buffer_pct=Decimal("0.10"),
+        )
+        manager = LiveRiskManager(config=config)
+
+        # Start of day
+        manager.reset_daily_tracking()
+        assert manager._start_of_day_equity is None
+
+        # First equity update
+        manager.track_daily_pnl(Decimal("10000"), {})
+        assert manager._start_of_day_equity == Decimal("10000")
+
+        # Normal trading
+        manager.pre_trade_validate(
+            symbol="BTC-USD",
+            side="buy",
+            quantity=Decimal("0.1"),
+            price=Decimal("50000"),
+            product=None,
+            equity=Decimal("10000"),
+            current_positions={},
+        )
+
+        # Record metrics
+        manager.append_risk_metrics(Decimal("10000"), {})
+        assert len(manager._risk_metrics) == 1
+
+        # Small loss - should not trigger
+        assert manager.track_daily_pnl(Decimal("9800"), {}) is False
+
+        # Large loss - should trigger
+        assert manager.track_daily_pnl(Decimal("9000"), {}) is True
+        assert manager.is_reduce_only_mode() is True
+
+        # End of day reset
+        manager.reset_daily_tracking()
+        assert manager.is_reduce_only_mode() is False
+
+    def test_volatility_triggers_reduce_only(self) -> None:
+        """Should trigger reduce-only mode on high volatility."""
+        config = MockConfig(volatility_threshold_pct=Decimal("0.05"))
+        manager = LiveRiskManager(config=config)
+
+        assert manager.is_reduce_only_mode() is False
+
+        closes = [
+            Decimal("100"),
+            Decimal("100"),
+            Decimal("100"),
+            Decimal("100"),
+            Decimal("200"),
+        ]
+
+        result = manager.check_volatility_circuit_breaker("BTC-USD", closes)
+
+        assert result.triggered is True
+        assert manager.is_reduce_only_mode() is True
+
+    def test_liquidation_buffer_triggers_reduce_only(self) -> None:
+        """Should trigger reduce-only for symbol when buffer low."""
+        config = MockConfig(min_liquidation_buffer_pct=Decimal("0.20"))
+        manager = LiveRiskManager(config=config)
+
+        # Buffer = |105 - 100| / 105 ≈ 4.76%, threshold 20%
+        triggered = manager.check_liquidation_buffer(
+            "BTC-PERP",
+            {"liquidation_price": 100, "mark": 105},
+            Decimal("1000"),
+        )
+
+        assert triggered is True
+        assert manager.positions["BTC-PERP"]["reduce_only"] is True
